+++ python/cudf/cudf/core/dataframe.py
@@ -1405,7 +1405,7 @@ class DataFrame(object):
             data, forceindex=forceindex, name=name
         )
 
-    def drop(self, labels, axis=None, errors="raise"):
+    def drop(self, labels=None, axis=None, columns=None, errors="raise"):
         """Drop column(s)
 
         Parameters
@@ -1414,6 +1414,7 @@ class DataFrame(object):
             Name of column(s) to be dropped.
         axis : {0 or 'index', 1 or 'columns'}, default 0
             Only axis=1 is currently supported.
+        columns: array of column names, the same as using labels and axis=1
         errors : {'ignore', 'raise'}, default 'raise'
             This parameter is currently ignored.
 
@@ -1443,15 +1444,26 @@ class DataFrame(object):
         3    3
         4    4
         """
-        if axis == 0:
+        if axis == 0 and labels is not None:
             raise NotImplementedError("Can only drop columns, not rows")
         if errors != "raise":
             raise NotImplementedError("errors= keyword not implemented")
+        if labels is None and columns is None:
+            raise ValueError(
+                "Need to specify at least one of 'labels' or 'columns'"
+            )
+        if labels is not None and columns is not None:
+            raise ValueError("Cannot specify both 'labels' and 'columns'")
+
+        if labels is not None:
+            target = labels
+        else:
+            target = columns
 
         columns = (
-            [labels]
-            if isinstance(labels, (str, numbers.Number))
-            else list(labels)
+            [target]
+            if isinstance(target, (str, numbers.Number))
+            else list(target)
         )
         outdf = self.copy()
         for c in columns:
@@ -3063,20 +3075,25 @@ class DataFrame(object):
 
         return output_frame
 
-    def isnull(self, **kwargs):
+    def isnull(self):
         """Identify missing values in a DataFrame.
         """
-        return self._apply_support_method("isnull", **kwargs)
+        return self._apply_support_method("isnull")
 
-    def isna(self, **kwargs):
+    def isna(self):
         """Identify missing values in a DataFrame. Alias for isnull.
         """
-        return self.isnull(**kwargs)
+        return self.isnull()
 
-    def notna(self, **kwargs):
+    def notna(self):
         """Identify non-missing values in a DataFrame.
         """
-        return self._apply_support_method("notna", **kwargs)
+        return self._apply_support_method("notna")
+
+    def notnull(self):
+        """Identify non-missing values in a DataFrame. Alias for notna.
+        """
+        return self.notna()
 
     def to_pandas(self):
         """